# platform/api/openapi/v1/horecaos-api.storefront.json
{
  "components" : {
    "schemas" : {
...
      "AddressFields" : {
        "properties" : {
          "apartment" : {
            "maxLength" : 32,
            "minLength" : 0,
            "type" : "string"
          },
          "city" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "district" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "entrance" : {
            "maxLength" : 32,
            "minLength" : 0,
            "type" : "string"
          },
          "floor" : {
            "maxLength" : 32,
            "minLength" : 0,
            "type" : "string"
          },
          "landmark" : {
            "maxLength" : 300,
            "minLength" : 0,
            "type" : "string"
          },
          "line1" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "line2" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "postalCode" : {
            "maxLength" : 32,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "SaveAddressRequest" : {
        "properties" : {
          "coordinateSource" : {
            "enum" : [ "NOT_GEOCODED", "LANDMARK_ONLY", "GEOCODER", "CUSTOMER_PIN", "OPERATOR_PIN", "LEGACY_UNSOURCED" ],
            "type" : "string"
          },
          "deliveryInstructions" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "fields" : {
            "$ref" : "#/components/schemas/AddressFields"
          },
          "label" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          },
          "latitude" : {
            "format" : "double",
            "maximum" : 90,
            "minimum" : -90,
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "maximum" : 180,
            "minimum" : -180,
            "type" : "number"
          }
        },
        "required" : [ "coordinateSource", "fields" ],
        "type" : "object"
      },
...
      "StorefrontCustomerControllerAddressResponse" : {
        "properties" : {
          "addressId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "coordinateSource" : {
            "enum" : [ "NOT_GEOCODED", "LANDMARK_ONLY", "GEOCODER", "CUSTOMER_PIN", "OPERATOR_PIN", "LEGACY_UNSOURCED" ],
            "type" : "string"
          },
          "deliveryInstructions" : {
            "type" : "string"
          },
          "fields" : {
            "$ref" : "#/components/schemas/AddressFields"
          },
          "label" : {
            "type" : "string"
          },
          "latitude" : {
            "format" : "double",
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "type" : "number"
          },
          "version" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "TermsView" : {
        "properties" : {
          "body" : {
            "type" : "string"
          },
          "isPlatformDefault" : {
            "type" : "boolean"
          },
          "locale" : {
            "type" : "string"
          },
          "version" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Keycloak access token issued for the horecaos-api audience",
        "scheme" : "bearer",
        "type" : "http"
      },
      "storefrontAppId" : {
        "description" : "The registered app's id. An identifier, not a secret: it appears in a browser bundle by design. Sent alongside the customer's session, never instead of it.",
        "in" : "header",
        "name" : "X-Storefront-App-Id",
        "type" : "apiKey"
      },
      "storefrontAppSecret" : {
        "description" : "A confidential client's secret. Server-side only; a public client has none and must not send one.",
        "in" : "header",
        "name" : "X-Storefront-App-Secret",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
...
    "title" : "HorecaOS Platform API",
    "version" : "v1",
    "x-horecaos-contract" : {
      "appIdHeader" : "X-Storefront-App-Id",
      "appSecretHeader" : "X-Storefront-App-Secret",
      "deprecationWindowMonths" : 12,
      "publishedPathsNeverRemovedWithin" : "v1"
    }
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/storefront/tenants/{tenantId}/brands/{brandId}/me/addresses/{addressId}" : {
      "delete" : {
        "description" : "Archived rather than deleted. The row is what a dispute about where an order went is answered from, and nothing depends on it staying readable to the customer: a cart and an order each hold their own copy, taken when the address was chosen, so an order in flight is unaffected. This is not an erasure request, which is a governed act over the whole account.",
        "operationId" : "removeAddress",
        "parameters" : [ {
          "in" : "path",
          "name" : "tenantId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "addressId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Remove one of the caller's own addresses",
        "tags" : [ "Customer self-service" ],
        "x-horecaos-surface" : "published"
      },
      "get" : {
        "description" : "Not found for an address that is somebody else's, is archived, or never existed. The three are one answer, because telling them apart is how an address id becomes probeable.",
        "operationId" : "address",
        "parameters" : [ {
          "in" : "path",
          "name" : "tenantId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "addressId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StorefrontCustomerControllerAddressResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "One of the caller's own saved addresses",
        "tags" : [ "Customer self-service" ],
        "x-horecaos-surface" : "published"
      },
      "put" : {
        "description" : "The whole address, never a field of it: the lines live inside one encrypted document, and the coordinate and its source have to move together. Requires If-Match, so a second tab loses loudly.",
        "operationId" : "updateAddress",
        "parameters" : [ {
          "in" : "path",
          "name" : "tenantId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "addressId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SaveAddressRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StorefrontCustomerControllerAddressResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Replace one of the caller's own addresses",
        "tags" : [ "Customer self-service" ],
        "x-horecaos-surface" : "published"
      }
    },
...
    "/api/v1/storefront/tenants/{tenantId}/brands/{brandId}/terms" : {
      "get" : {
        "description" : "The tenant's own published text when it has one for this language; otherwise the platform's brand-neutral default with brandName interpolated in. Cached briefly like the FAQ, because this changes on the order of weeks.",
        "operationId" : "current",
        "parameters" : [ {
          "in" : "path",
          "name" : "tenantId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "path",
          "name" : "brandId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "locale",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "brandName",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TermsView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "The document in force right now, in the requested language",
        "tags" : [ "Storefront terms of service" ],
        "x-horecaos-surface" : "published"
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ],
...
}